Safari-611.1.10/LayoutTests/platform/ios-13/fast/text/international/system-language/navigator-language/navigator-language-ru-expected.txt

FAIL navigator.language should be ru-ru. Was ru.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
